use KVC to support property type

+++ FishBindDemo-iOS/TestB.h
@@ -12,6 +12,6 @@
 
 @property (nonatomic, copy) NSString *bName;
 
-@property (nonatomic, assign) double ageB;
+@property (nonatomic, assign) NSInteger ageB;
 
 @end
